refactor(xtensa): split classic-ESP32 (LX6) glue into its own module
Mirror the esp32s3 split: move configure_xtensa_esp32, register_esp32_peripherals,
ESP32_PERIPHERALS, and attach_esp32_external_devices into system/xtensa/esp32.rs.
xtensa/mod.rs (1371 -> 672) now holds only the generic configure_xtensa + the
shared RamPeripheral + tests; re-exports esp32 so external paths are unchanged.

Symmetric structure: mod.rs (shared) + esp32.rs (LX6) + esp32s3.rs (S3). Pure
move — both feature sets + firmware_survival + register_coverage + brom green.
